TaskScheduler/requirements.txt
Django==3.2.16
django-extensions==3.2.1
djangorestframework==3.14.0
ipdb==0.13.13
ipython==8.12.0
dj-database-url==1.3.0
psycopg2==2.9.3
